New spell - Rest
Heals player to full at the start of next turn, but puts player to sleep for a turn.
This also hasthe remnants of an actual Kaclang spell that I tried to make, but that alone added 200 lines of code to the checkattack function of fightcontroller, doing each check for each set of commands whether a player was invincible, and too many things went wrong. Might comt back to it in the future, however.