Metal Slash working?
I think metal slash is working - I was originally going about it in a much more roundabout way by having cases for each scenario to replace text
Text is only replaced if metal slash is used, this was so much easier to work with.
Oh yeah, and Kaclang now prevents effects from being applied to the user the turn they're invincible.